# Improve `twilight_mention::fmt` performance by ~10%