--- a/mention/Cargo.toml
+++ b/mention/Cargo.toml
@@ -18,4 +18,10 @@ version = "0.4.1"
 twilight-model = { default-features = false, path = "../model" }
 
 [dev-dependencies]
+criterion = "0.3"
 static_assertions = { default-features = false, version = "1" }
+
+[[bench]]
+name = "fmt"
+harness = false
+path = "benches/fmt.rs"