Improve the performance of `twilight_mention::fmt` by around 10% by calling `Display::fmt` and `Formatter::write_str` directly instead of using `write!` and `format_args!`, which are notoriously ([1], [2]) slow.

Benchmarks:

```
format channel id       time:   [30.245 ns 30.317 ns 30.393 ns]
                        change: [-9.1152% -8.6720% -8.2652%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild

format emoji id         time:   [31.604 ns 31.629 ns 31.658 ns]
                        change: [-11.386% -10.974% -10.517%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  2 (2.00%) high mild
  5 (5.00%) high severe

format role id          time:   [31.130 ns 31.155 ns 31.186 ns]
                        change: [-11.698% -11.386% -11.081%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low severe
  3 (3.00%) high mild
  3 (3.00%) high severe

format user id          time:   [29.444 ns 29.473 ns 29.507 ns]
                        change: [-12.385% -12.045% -11.657%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  6 (6.00%) low severe
  2 (2.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe
```

[1]: https://github.com/rust-lang/rust/issues/76490
[2]: https://github.com/rust-lang/rust/issues/10761