@vivian 2021-06-18
Fails due to Rust 1.53 Clippy. #941 separately fixes them.